capture backtrace in throw statement
Before this commit program compiled by Clay printed backtrace in
assert function, but when exception is thrown/caught, exception
info is unavaiable. This is a problem when generic exception like
ValueMissing from maybe module is thrown, because virtually any
part of a program may throw this exception.

This commit changes throwValue function to capture backtrace and
store it next to exception. Captured backtrace and throw statement
location is available in the catch block as __exc_bt, __exc_file
and __exc_line variables.

Main function launcher now prints backtrace, and file and line of
throw statement on uncaught exception.

Unlike Java or C#, backtrace is not stored in exception object.
This seems to be more convenient in Clay, because Clay has no
inheritance, there is not base type to store backtrace in.

Backtrace capturing is a relatively fast operation (<1us in my
tests, see https://gist.github.com/3675622). But if this is too
slow for somebody, capturing can be turned off with compiler flag
-Dclay.DisableBacktraceOnThrow.

Unhandled `main` function exceptions is printed to stderr like this:

```
unhandled exception: MyException()
at ./tmp.clay:8
0   tmp                                 0x0001b878 __main__.main() clay + 632
1   tmp                                 0x0001b129 __operators__.callMain(Static[main], Int32, Pointer[Pointer[Int8]]) Int32 clay + 57
2   tmp                                 0x0001b0c2 main + 50
3   libdyld.dylib                       0x991eb725 start + 0
```