reorganize tests
Move all tests into subfolders.

* lang             -- language and compiler tests, including __primitives__
* lib-clay         -- stdlib tests
* lib-clay/core    -- tests of core.*
* example          -- tests that are example of what is Clay capable of
* example/shootout -- moved from test/shootout

This is necessary to quickly find out which parts of the project
are covered with tests.

The downside of this separation is that sometimes tests of similar
entities lies in different folders, like lang/floats and lib-clay/floats.

Patch contains nothing but file renames. No test files are modified
or removed.